[frontend] Move store persistence helpers to own file
