+++ frontend/src/store/modules/detailGraphStore.ts
@@ -21,86 +21,6 @@ export type DimensionDetailPoint = {
   dimension: Dimension
 }
 
-function hydrateDimension(it: Dimension) {
-  return new Dimension(it.benchmark, it.metric, it.unit, it.interpretation)
-}
-
-function hydrateDetailPoint(it: DetailDataPoint) {
-  return new DetailDataPoint(
-    it.hash,
-    it.parents,
-    it.author,
-    new Date(it.authorDate),
-    it.summary,
-    new CustomKeyEqualsMap(
-      it.values,
-      (first, second) =>
-        first.benchmark === second.benchmark && first.metric === second.metric
-    )
-  )
-}
-
-function deserializeDimensionDetailPoint(
-  point: DimensionDetailPoint | null
-): DimensionDetailPoint | null {
-  if (!point) {
-    return null
-  }
-  return {
-    dimension: hydrateDimension(point.dimension),
-    dataPoint: hydrateDetailPoint(point.dataPoint)
-  }
-}
-
-export function detailGraphStoreFromJson(json?: string): any {
-  if (!json) {
-    return {}
-  }
-
-  const parsedUnsafe = JSON.parse(json)
-  const parsed: DetailGraphStore = parsedUnsafe as DetailGraphStore
-  // Convert flat json to real object
-  parsedUnsafe._selectedDimensions = parsedUnsafe._selectedDimensions.map(
-    hydrateDimension
-  )
-  parsed.referenceDatapoint = deserializeDimensionDetailPoint(
-    parsed.referenceDatapoint
-  )
-  parsed.commitToCompare = deserializeDimensionDetailPoint(
-    parsed.commitToCompare
-  )
-  parsed.startTime = new Date(parsed.startTime)
-  parsed.endTime = new Date(parsed.endTime)
-
-  return parsed
-}
-
-function serializeDimensionDetailPoint(point: DimensionDetailPoint | null) {
-  if (!point) {
-    return null
-  }
-  const persistablePoint: any = Object.assign({}, point.dataPoint)
-  persistablePoint.values = Array.from(persistablePoint.values.entries())
-  return {
-    dataPoint: persistablePoint,
-    dimension: point.dimension
-  }
-}
-
-export function detailGraphStoreToJson(store: DetailGraphStore): string {
-  return JSON.stringify({
-    _selectedRepoId: (store as any)._selectedRepoId,
-    _selectedDimensions: (store as any)._selectedDimensions,
-    referenceDatapoint: serializeDimensionDetailPoint(store.referenceDatapoint),
-    startTime: store.startTime,
-    endTime: store.endTime,
-    duration: store.duration,
-    zoomStartValue: store.zoomStartValue,
-    zoomEndValue: store.zoomEndValue,
-    commitToCompare: serializeDimensionDetailPoint(store.commitToCompare)
-  })
-}
-
 export class DetailGraphStore extends VxModule {
   private _detailGraph: DetailDataPoint[] = []
   private _selectedRepoId: RepoId = ''
